requirements.txt
openai==1.63.2
python-dotenv==1.0.1